WIP - test for logResults
